Experiment with layouts and styles
Create a draft for layouts with interfaces, an abstract base layout
and a first implementating of a flexible linear layout.

Create a base UIComponent, and a simplified first styling system
based on colors only.

Delete outdated components.

This is experimentation only, some/all of this may never be seen in
the master branch.