Create interfaces for element layout, and a first implementation.
Create interfaces for layouts and elements to be affected by layouts.

Create an abstract LayoutBase implementing common layout functionality.

Create a "FlexLayout" based on the layout of the same name in CSS.

The current FlexLayout implementation is a messy first draft, and will
likely require future debugging and maybe optimization.

Finally, create a TestLayoutElement with which to test the FlexLayout.