Add margins, padding and content alignment for UI components
Margins and padding added this way are handled by the
UIComponent, and are not considered by the layout.

Also add a content alignment for elements whose content do not take
up the entire size of the element, such as the TextArea.

As a result of either of the above, different renderers (Text or
Background, at the moment) now render with different positions.
It becomes important that they be able to render at an offset from
the GameObject's transform. This is a given for the Duality sprite
renderer, but the duality text renderer does not allow this.

Copy the duality text renderer and make a small modification to allow
this.

Finally, update the TextArea test Scene to test these.