Add FormattedText-related properties to TextArea
Changes to the underlying FormattedText of a TextArea cannot be observed
and do not cause the layout to update. Add the relevant properties
as direct properties of TextArea.

This change reveals a bug in the way the TextArea computes its
preferred size - it adheres to the former max height and width
when wrapping. Remove the max size restriction while computing preferred
size.