Make adjustments to preferred size calculation
In order to help with supporting text wrapping, add the ability
for preferred size to take maximum size into account. This will only
apply when calculating content size after layout is complete, for now.

Add properties for Background position and size calculation, to help
simplify Content position and size calculation.

Text wrapping as is can cause the application to hang - this happens
if the TextArea width is small enough that no characters can be
rendered, but the height is unbounded. The default behaviour of the
FormattedText will be to check all 100,000,000 equally sized empty
lines in order to check if one could fit a glyph.

Ideally either this behaviour should be changed (either in Duality or
with a custom copy of FormattedText) but for now, as a simple fix,
add bound on the size of the TextArea that applies even when it would
otherwise be unbounded.