Create a Button UI element
Effectively just a UIComponent mouse listener. For now, have it
handle style, though this should change later.

Make a simplification to the UIContext so that the ContextHelper
does not have to be involved externally.

Allow the ColorBackground and background interface to allow its color
to be set. This is the extent of styling for now.