Include depth offsets as part of layout
Since all components are at the same z, depth offsets are the only
way of deciding what is at the front and visible/clickable.

Since higher depth offsets give the objects further back, and only
positive offsets are visible in screen space, this must be done
"backwards". The root layout has the highest offset and
child layouts and elements are placed closer and closer to zero.