Add TextButton
Up until now, the Button has been a UIComponent itself, with layout
information.  Only one UIElement is currently allowed per GameObject.

Changing this so that the button acts on the renderer
without regard for layout means the existing TextArea can be used
with it. A TextButton is then just a Button that requires a TextArea.