Delete previous components, styling and helpers
I'm going to try and take a different approach to stlying and templates
that will require a refactor across the board.

For now, delete everything from the old system.

They will still be present in history, and much of the old code will
be reintroduced in some fashion. (I certainly have no wish to try re-code
layout)