Introduce behaviours and styles
Create a BlueObject, which acts as the central hub for UI properties
in a GameObject. These draw values for properties from different sources,
including styles and templates. Create the Behaviour, which is a Component
that holds properties backed by the BlueObject.

Create some helper classes for notification and events.

Creat a specialized PropertyEditor for Behaviour Components.

There is a lot of experimenting going on here. Much of this code will be
refactored or deleted. Actually, the whole thing might still be deleted.