Fix Template save behaviour
Inspecting a template in the editor which has a property value set causes
the editor to think the template needs saving, before any editing has been done.

This is because the editor for a BlueProperty, on entering edit mode, always
sets the value in order to ensure that a value is set for that property.

To avoid this, provide the editor with a means to check if the property already
has a value.